+++ src/axom/inlet/Reader.hpp
@@ -17,6 +17,7 @@
 #include <functional>
 #include <string>
 #include <unordered_map>
+#include <unordered_set>
 #include <vector>
 
 #include "axom/inlet/Function.hpp"
@@ -247,6 +248,17 @@ public:
   virtual FunctionVariant getFunction(const std::string& id,
                                       const FunctionTag ret_type,
                                       const std::vector<FunctionTag>& arg_types) = 0;
+
+  /*!
+   *****************************************************************************
+   * \brief Retrieves all paths present in the input file
+   *
+   * \return The set of all paths/full names in the input file - this represents
+   * a full traversal of the input file "tree" and includes paths to array/dictionary
+   * entries
+   *****************************************************************************
+   */
+  virtual std::unordered_set<std::string> getAllNames() = 0;
 };
 
 }  // end namespace inlet